refactor(plugins): move GetPlugin to plugins/manager
